Move formatRequirements() and normalizeRequirements() into BaseAction
