+++ src/Composer/Action/RequirePackage.php
@@ -157,34 +157,4 @@ final class RequirePackage extends BaseAction
 
         return true;
     }
-
-    /**
-     * @param array $packages
-     *
-     * @return array
-     */
-    protected function formatRequirements(array $packages)
-    {
-        $requires = [];
-        $packages = $this->normalizeRequirements($packages);
-        foreach ($packages as $package) {
-            $requires[$package['name']] = $package['version'];
-        }
-
-        return $requires;
-    }
-
-    /**
-     * Parses a name/version pairs and returns an array of pairs.
-     *
-     * @param array $packages a set of package/version pairs separated by ":", "=" or " "
-     *
-     * @return array[] An array of arrays containing a name and (if provided) a version
-     */
-    protected function normalizeRequirements(array $packages)
-    {
-        $parser = new VersionParser();
-
-        return $parser->parseNameVersionPairs($packages);
-    }
 }
